관리자 통계, 화면깨짐 수정

+++ src/main/java/org/example/dao/OrderListDao.java
@@ -29,5 +29,5 @@ public interface OrderListDao {
 
     List<Map<String,Object>> getStat() throws Exception;
 
-    Map<String, Integer> getSumAvg(int i) throws Exception;
+    Map<String, Double> getSumAvg(int i) throws Exception;
 }
